requirements.txt
# NOTE: This file is kept for reference only.
# The canonical dependency specification is in pyproject.toml.
# Use `uv sync` or `pip install .` to install dependencies.

numpy>=2.0.0
scipy>=1.14.0
scikit-learn>=1.5.0
matplotlib>=3.9.0
umap-learn>=0.5.5
faiss-cpu>=1.9.0
numba>=0.60.0
llvmlite>=0.43.0
